# packages/contracts/openapi/monomarket-tickets.yaml
openapi: 3.0.3
info:
  title: MonoMarket Tickets API
  description: |
    Event marketplace platform API for ticket sales, management, and analytics.
    
    ## Features
    - Public marketplace for event discovery and checkout
    - Organizer panel for event and sales management
    - Director panel for platform administration
    - Payment processing with Conekta and Mercado Pago
    - Automated ticket generation with QR codes and PDF delivery
  version: 1.0.0
  contact:
    name: MonoMarket Tickets Team
servers:
  - url: http://localhost:3000/api
    description: Local development
  - url: https://api.monomarket-tickets.com/api
    description: Production
...
paths:
...
  /public/events:
    get:
      tags: [Public]
      summary: List all public events
      description: Get paginated list of published events
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of events
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /public/events/{id}:
    get:
      tags: [Public]
      summary: Get event details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Event details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Event not found

  # ========================================
  # CHECKOUT & PAYMENTS
  # ========================================
...
  /organizer/events:
    get:
      tags: [Organizer]
      summary: List organizer events
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
    post:
      tags: [Organizer]
      summary: Create new event
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventInput'
      responses:
        '201':
          description: Event created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
  /organizer/events/{id}:
    get:
      tags: [Organizer]
      summary: Get event details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Event details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
    put:
      tags: [Organizer]
      summary: Update event
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventInput'
      responses:
        '200':
          description: Event updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
    delete:
      tags: [Organizer]
      summary: Delete event
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Event deleted
...
  /tickets/validate/{qrCode}:
    get:
      tags: [Organizer]
      summary: Validate ticket by QR code
      description: Check ticket validity without checking in
      security:
        - bearerAuth: []
      parameters:
        - name: qrCode
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ticket validation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  qrCode:
                    type: string
                  status:
                    type: string
                    enum: [VALID, USED, CANCELLED]
                  usedAt:
                    type: string
                    format: date-time
                  event:
                    type: object
                    properties:
                      id:
                        type: string
                      title:
                        type: string
                      startDate:
                        type: string
                        format: date-time
                      venue:
                        type: string
                      address:
                        type: string
                      city:
                        type: string
                  buyer:
                    type: object
                    properties:
                      name:
                        type: string
                      email:
                        type: string
                  template:
                    type: object
                    properties:
                      name:
                        type: string
                  orderStatus:
                    type: string
        '404':
          description: Ticket not found
  /tickets/check-in/{qrCode}:
    post:
      tags: [Organizer]
      summary: Check in a ticket
      description: Mark ticket as USED and increment event attendance
      security:
        - bearerAuth: []
      parameters:
        - name: qrCode
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Check-in successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  ticket:
                    type: object
                    properties:
                      id:
                        type: string
                      qrCode:
                        type: string
                      status:
                        type: string
                      usedAt:
                        type: string
                        format: date-time
                      buyer:
                        type: object
                        properties:
                          name:
                            type: string
                          email:
                            type: string
                      template:
                        type: object
                        properties:
                          name:
                            type: string
                  event:
                    type: object
                    properties:
                      id:
                        type: string
                      title:
                        type: string
                      attendanceCount:
                        type: integer
        '400':
          description: Ticket already used or invalid
  /tickets/event/{eventId}/attendance:
    get:
      tags: [Organizer]
      summary: Get event attendance statistics
      security:
        - bearerAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Attendance statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  eventId:
                    type: string
                  eventTitle:
                    type: string
                  attendanceCount:
                    type: integer
                  totalTickets:
                    type: integer
                  percentageAttended:
                    type: number

  # ========================================

  # DIRECTOR ENDPOINTS
  # ========================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [organizer, director, admin]
        createdAt:
          type: string
          format: date-time
    Organizer:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        businessName:
          type: string
        status:
          type: string
          enum: [active, suspended, banned]
        feePlanId:
          type: string
        createdAt:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
    Event:
      type: object
      properties:
        id:
          type: string
        organizerId:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
        venue:
          type: string
        address:
          type: string
        city:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        coverImage:
          type: string
        status:
          type: string
          enum: [draft, published, cancelled]
        createdAt:
          type: string
          format: date-time
        organizer:
          $ref: '#/components/schemas/Organizer'
    EventInput:
      type: object
      required:
        - title
        - startDate
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        venue:
          type: string
        address:
          type: string
        city:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        coverImage:
          type: string
        status:
          type: string
          enum: [draft, published]
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer